+++ test/lambdas/test_koodistot_loader.py
